Add test to validate that the user name in cert must match

+++ run_demo.sh
@@ -1,9 +1,10 @@
 #!/bin/bash -ex
 
-type -p 
+test "$USER" = root && exit 5
 test -r sshd_config || exit 5
 rm -Rf keys work
 mkdir keys work
+workdir=$(pwd)/work
 
 myhost=$(hostname -f)
 tests_ok=()
@@ -12,8 +13,10 @@ tests_ok=()
 ssh-keygen -t dsa -N "" -C "SSH PKI CA Demo" -f keys/ca
 tests_ok+=(create-ca-key)
 
-# add CA key to known_hosts
+# add CA key to known_hosts and to authorized_keys
 echo "@cert-authority * $(ssh-keygen -y -f keys/ca)" >keys/ssh_known_hosts
+mkdir -p work/TRANSL/ssh-pki-demo/.ssh
+echo "cert-authority $(< keys/ca.pub)" >work/TRANSL/ssh-pki-demo/.ssh/authorized_keys
 
 # generate host key
 ssh-keygen -t dsa -N "" -C "Host key for $myhost" -f keys/host_key
@@ -23,17 +26,35 @@ tests_ok+=(create-host-key)
 ssh-keygen -s keys/ca -V +53w -I $myhost -h -n $myhost keys/host_key.pub
 tests_ok+=(sign-host-key)
 
+# generate user key for root
+ssh-keygen -t dsa -N "" -C "User key for root" -f keys/user_key_root
+tests_ok+=(create-user-root-key)
+
+# sign user key for root
+ssh-keygen -s keys/ca -V +53w -I root@$myhost -n root keys/user_key_root.pub
+tests_ok+=(sign-user-root-key)
+
+# generate user key for unpriv
+ssh-keygen -t dsa -N "" -C "User key for unpriv" -f keys/user_key_unpriv
+tests_ok+=(create-user-unpriv-key)
+
+# sign user key for root
+ssh-keygen -s keys/ca -V +53w -I unpriv@$myhost -n unpriv keys/user_key_unpriv.pub
+tests_ok+=(sign-user-unpriv-key)
+
 # seed work dir for installwatch
-mkdir -p work/TRANSL/etc/ssh work/TRANSL/ssh-pki-demo
-cp -v shadow passwd work/TRANSL/etc
-cp -v ssh_config sshd_config work/TRANSL/etc/ssh
-cp -rv keys/* work/TRANSL/etc/ssh/
+mkdir -p $workdir/TRANSL/etc/ssh 
+cp -v shadow passwd $workdir/TRANSL/etc
+# child processes spawned by sshd ignore fakeroot and installwatch, must set real path as fake root homedir for authorized_keys to work
+sed -e "s#/ssh-pki-demo#$workdir/TRANSL/ssh-pki-demo#" -i $workdir/TRANSL/etc/passwd
+cp -v ssh_config sshd_config $workdir/TRANSL/etc/ssh
+cp -rv keys/* $workdir/TRANSL/etc/ssh/
 
 function cleanup {
-	set +x
+	set +x +e
 	echo ; echo
 	echo "SSH PKI Demo Test Results:" ; echo
-	test "$daemon" && kill $daemon
+	test "$daemon" && kill -9 $daemon
 	for t in "${tests_ok[@]}" ; do
 		echo "Succeeded $t"
 	done
@@ -45,10 +66,9 @@ daemon=
 function start_daemon_run_ssh {
 	# $1 is label, remaining args are passed into eval and should return success
 	local label="$1" ; shift
-	local workdir=$(pwd)/work
 
 	# start ssh daemon
-	fakeroot installwatch -t -b -r $workdir /usr/sbin/sshd -ddd &>work/sshd.log &
+	fakeroot installwatch -t -b -r $workdir /usr/sbin/sshd -ddd &>$workdir/sshd.log &
 	daemon=$!
 
 	sleep 0.2
@@ -58,7 +78,7 @@ function start_daemon_run_ssh {
 	fi
 
 	# connect to our daemon and ask for date
-	if eval "fakeroot installwatch -t -r $workdir $@" 2>work/ssh.log ; then
+	if eval "fakeroot installwatch -t -r $workdir $@" 2>$workdir/ssh.log ; then
 		tests_ok+=("$label")
 		sleep 0.2
 		if kill -0 $daemon 2>/dev/null ; then
@@ -71,6 +91,19 @@ function start_daemon_run_ssh {
 	fi
 }
 
-start_daemon_run_ssh test-trusting-known-hosts-via-cert-and-login-with-password sshpass -pdemo ssh -vvv root@$myhost  date  +ssh-pki-demo_%c \| grep ssh-pki-demo
-start_daemon_run_ssh test-that-hostname-in-cert-must-match sshpass -pdemo ssh -vvv root@localhost  date  +ssh-pki-demo_%c \; test '$?' -gt 0
+start_daemon_run_ssh test-trusting-known-hosts-via-cert-and-login-with-password \
+	sshpass -pdemo ssh -vvv root@$myhost  date  +ssh-pki-demo_%c \| grep ssh-pki-demo
+
+start_daemon_run_ssh test-that-hostname-in-cert-must-match-target-host \
+	sshpass -pdemo ssh -vvv root@localhost  date  +ssh-pki-demo_%c \; test '$?' -gt 0
+
+start_daemon_run_ssh test-login-with-root-key-trusted-by-cert \
+	ssh -vvv -o PreferredAuthentications=publickey -i /etc/ssh/user_key_root root@$myhost  date  +ssh-pki-demo_%c \| grep ssh-pki-demo \
+	"&&" grep Accepted.certificate.ID $workdir/sshd.log
+
+start_daemon_run_ssh test-that-username-in-cert-must-match-target-user \
+	ssh -vvv -o PreferredAuthentications=publickey -i /etc/ssh/user_key_unpriv root@$myhost  date  +ssh-pki-demo_%c \| grep ssh-pki-demo \
+	";" grep Permission.denied $workdir/ssh.log
+
+
 tests_ok+=("in running all tests, congratulations!")
